feat(agents): replace job dependencies with data-based requirements
Introduce DataRequirement class for more robust dependency checking.
Instead of verifying that a prior job ran, agents now check that the
actual data they need exists in the database with sufficient recency.

Changes:
- Add DataRequirement class with table, min_rows, max_age_days checks
- Update IngestionJob to support data_requirements parameter
- Refactor SignalScientist to use data requirements (features table)
- Refactor FeatureComputationJob to use data requirements (prices table)
- Add type annotations across data ingestion and validation modules
- Move completed vectorbt-pro plans to docs/reports/
- Add ML Quality Sentinel agent plan for next implementation
- Update tests for new data requirement approach

+++ hrp/research/metrics.py
@@ -83,7 +83,7 @@ def _calculate_cagr(returns: pd.Series, periods_per_year: int) -> float:
     n_years = len(returns) / periods_per_year
     if n_years <= 0 or total_return <= 0:
         return 0.0
-    return total_return ** (1 / n_years) - 1
+    return float(total_return ** (1 / n_years) - 1)
 
 
 def _sharpe_ratio(excess_returns: pd.Series, periods_per_year: int) -> float:
@@ -92,7 +92,7 @@ def _sharpe_ratio(excess_returns: pd.Series, periods_per_year: int) -> float:
     # Handle near-zero volatility (floating point precision)
     if std < 1e-10:
         return 0.0
-    return excess_returns.mean() / std * np.sqrt(periods_per_year)
+    return float(excess_returns.mean() / std * np.sqrt(periods_per_year))
 
 
 def _sortino_ratio(returns: pd.Series, risk_free_rate: float, periods_per_year: int) -> float:
@@ -101,7 +101,7 @@ def _sortino_ratio(returns: pd.Series, risk_free_rate: float, periods_per_year: 
     downside = _downside_volatility(returns, periods_per_year)
     if downside == 0:
         return 0.0
-    return excess.mean() * periods_per_year / downside
+    return float(excess.mean() * periods_per_year / downside)
 
 
 def _downside_volatility(returns: pd.Series, periods_per_year: int) -> float:
@@ -113,7 +113,7 @@ def _downside_volatility(returns: pd.Series, periods_per_year: int) -> float:
     # Handle near-zero volatility (floating point precision)
     if std < 1e-10:
         return 0.0
-    return std * np.sqrt(periods_per_year)
+    return float(std * np.sqrt(periods_per_year))
 
 
 def _max_drawdown(returns: pd.Series) -> float:
@@ -124,7 +124,7 @@ def _max_drawdown(returns: pd.Series) -> float:
     cumulative = pd.concat([pd.Series([1.0]), cumulative]).reset_index(drop=True)
     running_max = cumulative.cummax()
     drawdown = (cumulative - running_max) / running_max
-    return drawdown.min()
+    return float(drawdown.min())
 
 
 def _calculate_alpha_beta(returns: pd.Series, benchmark: pd.Series, periods_per_year: int) -> tuple[float, float]:
